refactor: separate diagnostics from scientific results
